Moved components to their correct component folders.
