+++ src/app/components/information/Dialog.tsx
